2020/03/13 02:15 Improve Vector class
Modify the constructor so that if a single integer is sent,
it produces a vector of that dimension with all zeros,
but if a sequence is provided, coordinates'll be based on that sequence.
For example, Vector([4, 7, 5]).

Implement the __sub__ method.

Implement the __neg__ method.

The syntax [5, 3, 10, −2, 1] + u is now legal.

Implement the __mul__ and __rmul methods. For example, u * 3, u * v.